+++ packages/web-visualization/src/chart/PeriodSelector.tsx
@@ -124,10 +124,10 @@ export const LiveTabLabel: LiveTabLabelComponent = memo(
   ),
 );
 
-// Custom tab component with primary color for active state
+// Set font
 const PeriodSelectorTab: TabComponent = memo(
   forwardRef((props: SegmentedTabProps, ref: React.ForwardedRef<HTMLButtonElement>) => (
-    <SegmentedTab ref={ref} activeColor="fgPrimary" font="label1" {...props} />
+    <SegmentedTab ref={ref} font="label1" {...props} />
   )),
 );
 
@@ -143,6 +143,7 @@ export const PeriodSelector = memo(
       {
         background = 'transparent',
         activeBackground = 'bgPrimaryWash',
+        activeColor = 'fgPrimary',
         width = '100%',
         justifyContent = 'space-between',
         TabComponent = PeriodSelectorTab,
@@ -160,6 +161,7 @@ export const PeriodSelector = memo(
         TabComponent={TabComponent}
         TabsActiveIndicatorComponent={TabsActiveIndicatorComponent}
         activeBackground={activeBackground}
+        activeColor={activeColor}
         background={background}
         className={cx(className, classNames?.root)}
         classNames={{