+++ packages/web/src/__stories__/componentConfigStickerSheet/customComponentConfig.tsx
@@ -134,4 +134,10 @@ export const customComponentConfig: ComponentConfig = {
       ...(spacingVariant === 'normal' ? { minHeight: 36 } : {}),
     };
   },
+
+  Tabs: {
+    activeColor: 'fg',
+    color: 'fgMuted',
+    activeBackground: 'fg',
+  },
 };