+++ packages/web/src/tabs/__stories__/SegmentedTabs.stories.tsx
@@ -244,6 +244,14 @@ export const All = () => {
         title="Icon Labels"
         width="fit-content"
       />
+      <SegmentedTabsExample
+        activeBackground="bgPositiveWash"
+        activeColor="bgPositive"
+        color="fgMuted"
+        defaultActiveTab={basicSegments[0]}
+        tabs={basicSegments}
+        title="Label Colors"
+      />
     </VStack>
   );
 };