+++ packages/mcp-server/CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 8.70.0 ((5/8/2026, 02:16 PM PST))
+
+This is an artificial version bump with no new change.
+
 ## 8.69.1 ((5/6/2026, 10:08 AM PST))
 
 This is an artificial version bump with no new change.